+++ libft/ft_atoi.c
@@ -6,14 +6,14 @@
 /*   By: kcisse <kcisse@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 16:11:25 by kcisse            #+#    #+#             */
-/*   Updated: 2024/05/30 14:41:04 by kcisse           ###   ########.fr       */
+/*   Updated: 2025/04/16 16:58:57 by kcisse           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	get_res(const char *str)
+long	get_res(const char *str)
 {
-	int	res;
-	int	i;
+	long	res;
+	int		i;
 
 	res = 0;
 	i = 0;
@@ -26,12 +26,12 @@ int	get_res(const char *str)
 }
 
 // transform char to int
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	res;
-	int	plus_sign;
-	int	minus_sign;
+	long	i;
+	long	res;
+	int		plus_sign;
+	int		minus_sign;
 
 	i = 0;
 	res = 0;
